pyproject.toml
[build-system]
requires = ["setuptools>=61.0"] # Minimum version of setuptools
build-backend = "setuptools.build_meta"

[project]
name = "debentures-dot-com" # This is the name users will `pip install`
version = "0.1.0" # Start with 0.1.0 and increment for releases
authors = [
    { name="Your Name", email="your.email@example.com" },
]
description = "A Python client for scraping debenture data from debentures.com"
readme = "README.md"
requires-python = ">=3.8" # Specify your Python version compatibility
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # Or your chosen license
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha", # Or 4 - Beta, 5 - Production/Stable
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests",
    "pandas",
    "beautifulsoup4",
    "python-dateutil", # This is what `dateutil.parser` comes from
]

[project.urls]
"Homepage" = "https://github.com/gtazevedo/debentures_dot_com"
"Bug Tracker" = "https://github.com/gtazevedo/debentures_dot_com/issues"

# If you have a 'src' layout, tell setuptools where to find the package
[tool.setuptools.packages.find]
where = ["src"]